requirements.txt
django==1.11.2
django-revproxy==0.9.13
django-jet==1.0.6
Pillow==4.1.1
uwsgi==2.0.15
